+++ csvkit/utilities/in2csv.py
@@ -3,6 +3,8 @@
 import agate
 import agatedbf  # noqa
 import agateexcel  # noqa
+import openpyxl
+import xlrd
 
 from csvkit import convert
 from csvkit.convert.fixed import fixed2csv
@@ -26,8 +28,10 @@ class In2CSV(CSVKitUtility):
                                     help='Specifies a CSV-formatted schema file for converting fixed-width files.  See documentation for details.')
         self.argparser.add_argument('-k', '--key', dest='key',
                                     help='Specifies a top-level key to use look within for a list of objects to be converted when processing JSON.')
+        self.argparser.add_argument('-n', '--names', dest='names_only', action='store_true',
+                                    help='Display sheet names from the input Excel file.')
         self.argparser.add_argument('--sheet', dest='sheet',
-                                    help='The name of the XLSX sheet to operate on.')
+                                    help='The name of the Excel sheet to operate on.')
         self.argparser.add_argument('-y', '--snifflimit', dest='sniff_limit', type=int,
                                     help='Limit CSV dialect sniffing to the specified number of bytes. Specify "0" to disable sniffing entirely.')
         self.argparser.add_argument('--no-inference', dest='no_inference', action='store_true',
@@ -50,6 +54,7 @@ class In2CSV(CSVKitUtility):
             if not filetype:
                 self.argparser.error('Unable to automatically determine the format of the input file. Try specifying a format with --format.')
 
+        # Buffer standard input if the input file is in CSV format or if performing type inference.
         self.buffers_input = filetype == 'csv' or not self.args.no_inference
 
         # Set the input file.
@@ -57,6 +62,20 @@ class In2CSV(CSVKitUtility):
             self.input_file = open(self.args.input_path, 'rb')
         else:
             self.input_file = self._open_input_file(self.args.input_path)
+
+        if self.args.names_only:
+            sheet_names = None
+            if filetype == 'xls':
+                sheet_names = xlrd.open_workbook(file_contents=self.input_file.read()).sheet_names()
+            elif filetype == 'xlsx':
+                sheet_names = openpyxl.load_workbook(self.input_file, read_only=True, data_only=True).sheetnames
+            if sheet_names:
+                for name in sheet_names:
+                    self.output_file.write('%s\n' % name)
+            else:
+                self.argparser.error('You cannot use the -n or --names options with non-Excel files.')
+            self.input_file.close()
+            return
 
         # Set the reader's arguments.
         kwargs = {}
@@ -71,14 +90,11 @@ class In2CSV(CSVKitUtility):
 
         if filetype == 'csv':
             kwargs.update(self.reader_kwargs)
-            # Streaming CSV musn't set sniff_limit, but non-streaming should.
-            if not self.args.no_inference:
-                kwargs['sniff_limit'] = self.args.sniff_limit
-            if self.args.no_header_row:
-                kwargs['header'] = False
-        elif self.args.no_inference:
-            # Streaming CSV musn't set column_types, but other formats should.
-            kwargs['column_types'] = agate.TypeTester(limit=0)
+            kwargs['sniff_limit'] = self.args.sniff_limit
+            kwargs['header'] = not self.args.no_header_row
+
+        if filetype != 'dbf':
+            kwargs['column_types'] = self.get_column_types()
 
         # Convert the file.
         if filetype == 'csv' and self.args.no_inference: