+++ CHANGELOG.rst
@@ -2,6 +2,7 @@
 -----
 
 * :doc:`/scripts/csvjoin` supports ``--snifflimit`` and ``--no-inference``.
+* :doc:`/scripts/in2csv` now supports a :code:`--names` flag to print Excel sheet names.
 
 1.0.1 - December 29, 2016
 -------------------------